+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `greyscale` filter for converting to 8-bit greyscale
 - Add `tint` filter for tinting with a provided colour
 - Add `lqip` helper for generating a low quality image placeholder. Thanks for the suggestion [@lukew-cogapp](https://github.com/dyerc/craft-flux/issues/19)
+- Add support for Avif and automatically serving Avif if accepted by client browser
 
 ## 5.1.1 - 2026-01-07